Move general_tests.rs into mod.rs
